soapy: set antenna in auto mode during setup stream procedure
SRS eNB application sets antennas only in case of parameters of these antennas was specified in the configuration file and forwarded via device args.
Parameters of antennas are band dependant, so these parameters should be changed each time we change the band in the configuration.
Luckily we have AUTO mode, in this mode, the proper antennas parameters will be set automatically depending on the used band.